day 2 done

+++ springbootexamples/bankapp03/src/main/java/com/bankapp/service/AccountService.java
@@ -1,5 +1,6 @@
 package com.bankapp.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.bankapp.entity.Account;
@@ -7,7 +8,7 @@ import com.bankapp.entity.Account;
 public interface AccountService {
 	public List<Account> getAll();
     public Account getById(int id);
-    public void transfer(int fromId, int toId, double amount);
-    public void deposit(int fromId, double amount);
-    public void withdraw(int fromId, double amount);
+    public void transfer(int fromId, int toId, BigDecimal amount);
+    public void deposit(int fromId, BigDecimal amount);
+    public void withdraw(int fromId, BigDecimal amount);
 }
